Made debugging for guilds more useful
Currently working on finding out when changes are made to a guild.
Whenever someone logs in or our I want to update a stored array of valid
players to avenge. To do this I need to know who is in the urgent dawn
guild and when they are in it.